+++ source_gen/source_gen/source_gen/lib/source_gen.dart
@@ -2,7 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/builder.dart' show defaultFileHeader, LibraryBuilder, PartBuilder;
+export 'src/builder.dart'
+    show defaultFileHeader, LibraryBuilder, PartBuilder, SharedPartBuilder;
 export 'src/constants/reader.dart' show ConstantReader;
 export 'src/constants/revive.dart' show Revivable;
 export 'src/generator.dart' show Generator, InvalidGenerationSourceError;