+++ source_gen/source_gen/source_gen/lib/src/builder.dart
@@ -28,6 +28,12 @@ class _Builder extends Builder {
   /// Whether to emit a standalone (non-`part`) file in this builder.
   final bool _isStandalone;
 
+  /// Whether to include `part of` in the output.
+  ///
+  /// This allows the easy merging of outputs without multiple `part of`
+  /// statements.
+  final bool _outputPartOf;
+
   final String _header;
 
   @override
@@ -39,14 +45,16 @@ class _Builder extends Builder {
       String generatedExtension = '.g.dart',
       List<String> additionalOutputExtensions = const [],
       bool isStandalone = false,
-      String header})
+      String header,
+      bool outputPartOf = true})
       : _generatedExtension = generatedExtension,
         buildExtensions = {
           '.dart': [generatedExtension]..addAll(additionalOutputExtensions)
         },
         _isStandalone = isStandalone,
         formatOutput = formatOutput ?? _formatter.format,
-        _header = (header ?? defaultFileHeader).trim() {
+        _header = (header ?? defaultFileHeader).trim(),
+        _outputPartOf = outputPartOf {
     if (_generatedExtension == null) {
       throw new ArgumentError.notNull('generatedExtension');
     }
@@ -104,7 +112,7 @@ class _Builder extends Builder {
         log.warning('Missing "part \'$part\';".');
       }
       contentBuffer.writeln();
-      contentBuffer.writeln('part of $name;');
+      if (_outputPartOf) contentBuffer.writeln('part of $name;');
     }
 
     for (var item in generatedOutputs) {
@@ -145,13 +153,49 @@ class _Builder extends Builder {
 }
 
 /// A [Builder] which generates `part of` files.
+///
+/// Generated files will be prefixed with a `partId` to ensure multiple
+/// [SharedPartBuilder]s can produce non conflicting `part of` files.
+class SharedPartBuilder extends _Builder {
+  /// Wrap [generators] as a [Builder] that generates `part of` files.
+  ///
+  /// [partId] indicates what files will be created for each `.dart`
+  /// input. This extension should be unique as to not conflict with other
+  /// [SharedPartBuilder]s. The resulting file will be of the form
+  /// `<generatedExtension>.g.part`. If any generator in [generators] will
+  /// create additional outputs through the [BuildStep] they should be indicated
+  /// in [additionalOutputExtensions].
+  ///
+  /// [formatOutput] is called to format the generated code. Defaults to
+  /// [DartFormatter.format].
+  SharedPartBuilder(List<Generator> generators, String partId,
+      {String formatOutput(String code),
+      List<String> additionalOutputExtensions = const []})
+      : super(generators,
+            formatOutput: formatOutput,
+            generatedExtension: '.$partId.g.part',
+            additionalOutputExtensions: additionalOutputExtensions,
+            outputPartOf: false) {
+    if (!_partIdRegExp.hasMatch(partId)) {
+      throw new ArgumentError.value(
+          partId,
+          'partId',
+          '`partId` can only contain letters, numbers, `_` and `.`. '
+          'It cannot start or end with `.`.');
+    }
+  }
+}
+
+/// A [Builder] which generates `part of` files.
 class PartBuilder extends _Builder {
   /// Wrap [generators] as a [Builder] that generates `part of` files.
   ///
-  /// [generatedExtension] indicates what files will be created for each `.dart`
-  /// input. Defaults to `.g.dart`. If any generator in [generators] will create
-  /// additional outputs through the [BuildStep] they should be indicated in
-  /// [additionalOutputExtensions].
+  /// [generatedExtension] indiciates what files will be created for each
+  /// `.dart` input. The [generatedExtension] should *not* be `.g.dart`. If you
+  /// wish to produce `.g.dart` files please use [SharedPartBuilder].
+  ///
+  /// If any generator in [generators] will create additional outputs through
+  /// the [BuildStep] they should be indicated in [additionalOutputExtensions].
   ///
   /// [formatOutput] is called to format the generated code. Defaults to
   /// [DartFormatter.format].
@@ -159,9 +203,8 @@ class PartBuilder extends _Builder {
   /// [header] is used to specify the content at the top of each generated file.
   /// If `null`, the content of [defaultFileHeader] is used.
   /// If [header] is an empty `String` no header is added.
-  PartBuilder(List<Generator> generators,
+  PartBuilder(List<Generator> generators, String generatedExtension,
       {String formatOutput(String code),
-      String generatedExtension = '.g.dart',
       List<String> additionalOutputExtensions = const [],
       String header})
       : super(generators,
@@ -234,3 +277,7 @@ final _formatter = new DartFormatter();
 const defaultFileHeader = '// GENERATED CODE - DO NOT MODIFY BY HAND';
 
 final _headerLine = '// '.padRight(77, '*');
+
+const partIdRegExpLiteral = r'[A-Za-z_\d]+';
+
+final _partIdRegExp = new RegExp('^$partIdRegExpLiteral\$');